feat(wordbank): fold C2 variant into production — click-to-select with label-row edit button
- Remove prototype scaffolding (PrototypeSwitcher, C1 variant)
- Fold C2 directly into WordBankEntryDetail:
  - Click a field to select (subtle accent highlight + ring)
  - Edit button fades/slides in on the label row — zero layout shift
  - Click outside the fields container to deselect
  - Escape key also deselects
- Keep the 150ms opacity+translateY animation on the Edit button
